BI-4121 Split api interface into own sdk/objections folder to make it easier to migrate to real sdk
