+++ src/services/objections.api.service.ts
@@ -1,80 +1,6 @@
-import { AxiosError, AxiosRequestConfig, AxiosResponse, Method } from "axios";
-import axios from "axios";
+
+import * as objectionsSdk from "../sdk/objections";
 import logger from "../utils/logger";
-import { INTERNAL_API_URL } from "../utils/properties";
-
-const HTTP_POST: Method = "post";
-
-/**
- * Api Error
- * @interface
- */
-export interface ApiError {
-  data: any;
-  message: string;
-  status: number;
-}
-
-/**
- * An axios config with common elements for API calls.
- * @param {Method} httpMethod the http method to use eg. get, post
- * @param {string} url of the api endpoint
- * @param {string} token Bearer token for API call
- */
-const getAxiosRequestConfig = (httpMethod: Method, url: string, bearerToken: string): AxiosRequestConfig => {
-  return {
-    headers: {
-      Accept: "application/json",
-      Authorization: "Bearer " + bearerToken,
-    },
-    method: httpMethod,
-    proxy: false,
-    url,
-  };
-};
-
-/**
- * Call the API using the supplied Axios request config.
- * @param {AxiosRequestConfig} axios request config
- * @returns {Promise<AxiosResponse>} the api response
- * @throws {ApiError}
- */
-const makeAPICall = async (config: AxiosRequestConfig): Promise<AxiosResponse> => {
-  try {
-    logger.debug(`Calling ${config.method} ${config.url} with data ${config.data}`);
-    const axiosResponse: AxiosResponse = await axios.request<any>(config);
-    logger.debug(`data returned from axios api call : ${JSON.stringify(axiosResponse.data, null, 2)}`);
-    return axiosResponse;
-  } catch (err) {
-    logger.error(`ERROR calling API ${JSON.stringify(err, null, 2)}`);
-    const axiosError = err as AxiosError;
-    const { response, message } = axiosError;
-    throw {
-      data: response ? response.data.errors : [],
-      message,
-      status: response ? response.status : -1,
-    } as ApiError;
-  }
-};
-
-/**
- * Status of Objection
- * @enum
- */
-export enum ObjectionStatus {
-  OPEN,
-  SUBMITTED,
-  PROCESSED,
-}
-
-/**
- * Data structure for patching an Objection
- * @interface
- */
-export interface ObjectionPatch {
-  reason?: string;
-  status?: ObjectionStatus;
-}
 
 /**
  * Create a new objection for the given company.
@@ -86,25 +12,34 @@ export interface ObjectionPatch {
  * @throws {ApiError}
  */
 export const createNewObjection = async (companyNumber: string, token: string): Promise<string> => {
-  logger.info(`Creating a new objection for company number ${companyNumber}`);
-
-  const createNewObjectionUrl = `${INTERNAL_API_URL}/company/${companyNumber}/strike-off-objections/`;
-  const axiosConfig: AxiosRequestConfig = getAxiosRequestConfig(HTTP_POST, createNewObjectionUrl, token);
-
-  return (await makeAPICall(axiosConfig)).data.id as string;
+  return await objectionsSdk.createNewObjection(companyNumber, token);
 };
 
 /**
- * Patch an objection for the given company.
+ * Update an objection reason for the given company.
  *
- * @param companyNumber the company number
- * @param token the bearer security token to use to call the api
- *
+ * @param {string} objectionId the id of the objection
+ * @param {string} token the bearer security token to use to call the api
+ * @param {string} reason
  */
-export const patchObjection = (companyNumber: string, token: string, patch: ObjectionPatch) => {
+export const updateObjectionReason = (objectionId: string, token: string, reason: string) => {
 
-  logger.debug(`Patching an objection for company number ${companyNumber}`);
+  logger.info(`Updating objection reason for objectionId ${objectionId}`);
 
-  // TODO Call the actual Objections API when end-point is implemented. Covered by JIRA
+  // TODO Call the Objections SDK. Covered by JIRA
+  //      sub-task BI-4143
+};
+
+/**
+ * Update an objection status for the given company.
+ *
+ * @param {string} objectionId the id of the objection
+ * @param {string} token the bearer security token to use to call the api
+ */
+export const updateObjectionStatusToSubmitted = (objectionId: string, token: string) => {
+
+ logger.info(`Updating objection status to submitted for objectionId ${objectionId}`);
+
+  // TODO Call the Objections SDK. Covered by JIRA
   //      sub-task BI-4143
 };
