+++ test/services/objections.api.service.spec.unit.ts
@@ -1,68 +1,33 @@
-jest.mock("axios");
+jest.mock("../../src/sdk/objections");
 
-import axios, { AxiosError, AxiosResponse } from "axios";
-import { ApiError, createNewObjection, ObjectionPatch, ObjectionStatus, patchObjection } from "../../src/services/objections.api.service";
+import * as objectionsSdk from "../../src/sdk/objections";
+import * as objectionsApiService from "../../src/services/objections.api.service";
 
-const mockAxiosRequest = axios.request as jest.Mock;
+const mockCreateNewObjection = objectionsSdk.createNewObjection as jest.Mock;
 
 const ACCESS_TOKEN = "KGGGUYUYJHHVK1234";
 const COMPANY_NUMBER = "00006400";
 
 describe("objections API service unit tests", () => {
 
-  beforeEach(() => {
-    mockAxiosRequest.mockReset();
-  });
-
   it("returns an id when a new objection is created", async () => {
     const NEW_OBJECTION_ID = "7687kjh-33kjkjkjh-hjgh435";
-    mockAxiosRequest.mockResolvedValueOnce({
-      data: {
-        id: NEW_OBJECTION_ID,
-      },
-    });
-    const objectionId: string = await createNewObjection(COMPANY_NUMBER, ACCESS_TOKEN);
+    mockCreateNewObjection.mockResolvedValueOnce(NEW_OBJECTION_ID);
+
+    const objectionId: string = await objectionsApiService.createNewObjection(COMPANY_NUMBER, ACCESS_TOKEN);
 
     expect(objectionId).toBeDefined();
     expect(typeof objectionId).toBe("string");
     expect(objectionId).toStrictEqual(NEW_OBJECTION_ID);
   });
 
-  it("throws an ApiError from createNewObjection when an error occurs calling the api", async () => {
-    const axiosError: AxiosError = {
-      message: "Not Found",
-      response: {
-        data: {
-          errors: ["missing company"],
-        },
-        status: 404,
-      } as AxiosResponse,
-    } as AxiosError;
-
-    mockAxiosRequest.mockRejectedValueOnce(axiosError);
-
-    let thrownError: ApiError;
-    try {
-      await createNewObjection(COMPANY_NUMBER, ACCESS_TOKEN);
-    } catch (e) {
-      thrownError = e;
-    }
-
-    const expectedApiError: ApiError = {
-      data: ["missing company"],
-      message: "Not Found",
-      status: 404,
-    };
-
-    expect(thrownError).toStrictEqual(expectedApiError);
+  it("returns undefined when updating an objection reason", () => {
+    const patchResult = objectionsApiService.updateObjectionReason(COMPANY_NUMBER, ACCESS_TOKEN, "reason");
+    expect(patchResult).toBeUndefined();
   });
 
-  it("returns undefined when patching an objection", () => {
-    const patch: ObjectionPatch = {
-      reason: "some reason or other",
-      status: ObjectionStatus.SUBMITTED,
-    };
-    const patchResult = patchObjection(COMPANY_NUMBER, ACCESS_TOKEN, patch);
+  it("returns undefined when updating an objection status to submitted", () => {
+    const patchResult = objectionsApiService.updateObjectionStatusToSubmitted(COMPANY_NUMBER, ACCESS_TOKEN);
     expect(patchResult).toBeUndefined();
   });
 });
